work on errors, content, and card controller
