+++ card/lib/card/auth.rb
@@ -1,128 +1,20 @@
 # -*- encoding : utf-8 -*-
+require_dependency "card/auth/permissions"
+require_dependency "card/auth/proxy"
+require_dependency "card/auth/setup"
 
 class Card
   # Singleton methods for authentication. Manages current user,
   # "as" user, and password verification.
   module Auth
-    extend Card::Auth::Permissions
-    extend Card::Auth::Proxy
-    extend Card::Auth::Setup
+    extend Permissions
+    extend Proxy
+    extend Setup
+    extend Current
 
     @as_card = @as_id = @current_id = @current = nil
 
     class << self
-      # set current user in process and session
-      def signin signin_id
-        self.current_id = signin_id
-        session[:user] = signin_id if session
-      end
-
-      # current user is not anonymous
-      # @return [true/false]
-      def signed_in?
-        current_id != Card::AnonymousID
-      end
-
-      # id of current user card.
-      # @return [Integer]
-      def current_id
-        @current_id ||= Card::AnonymousID
-      end
-
-      # current accounted card (must have +\*account)
-      # @return [Card]
-      def current
-        if @current && @current.id == current_id
-          @current
-        else
-          @current = Card[current_id]
-        end
-      end
-
-      # set the id of the current user.
-      def current_id= card_id
-        @current = @as_id = @as_card = nil
-        card_id = card_id.to_i if card_id.present?
-        @current_id = card_id
-      end
-
-      # set current user from email or id
-      # @return [Integer]
-      def current= mark
-        self.current_id =
-          if mark.to_s =~ /@/
-            account = Auth.find_account_by_email mark
-            account && account.active? ? account.left_id : Card::AnonymousID
-          else
-            mark
-          end
-      end
-
-      # get session object from Env
-      # return [Session]
-      def session
-        Card::Env[:session]
-      end
-
-      # get :user id from session and set Auth.current_id
-      def set_current_from_session
-        self.current_id =
-          if session
-            if (card_id = session[:user]) && Card.exists?(card_id)
-              card_id
-            else
-              session[:user] = nil
-            end
-          end
-        current_id
-      end
-
-      # set the current user based on token
-      def set_current_from_token token, current=nil
-        account = find_account_by_token token
-        if account && account.validate_token!(token)
-          unless current && always_ok_usr_id?(account.left_id)
-            # can override current only if admin
-            current = account.left_id
-          end
-          self.current = current
-        elsif Env.params[:live_token]
-          true
-          # Used for activations and resets.
-          # Continue as anonymous and address problem later
-        else
-          false
-        end
-      end
-
-      # find +\*account card by +\*token card
-      # @param token [String]
-      # @return [+*account card, nil]
-      def find_account_by_token token
-        find_account_by "token", Card::TokenID, token.strip
-      end
-
-      # find +\*account card by +\*email card
-      # @param email [String]
-      # @return [+*account card, nil]
-      def find_account_by_email email
-        find_account_by "email", Card::EmailID, email.strip.downcase
-      end
-
-      # general pattern for finding +\*account card based on field cards
-      # @param fieldname [String] right name of field card (for WQL comment)
-      # @param field_id [Integer] card id of field's simple card
-      # @param value [String] content of field
-      # @return [+*account card, nil]
-      def find_account_by fieldname, field_id, value
-        Auth.as_bot do
-          Card.search({ right_id: Card::AccountID,
-                        right_plus: [{ id: field_id },
-                                     { content: value }] },
-                      "find +*account for #{fieldname} (#{value})").first
-        end
-      end
-
       # authenticate a user by their login name and unencrypted password.
       # @param email [String]
       # @param password [String]
